+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Check out the [Upgrade Guide](docs/upgrade_v8.md).
 
 ### Changed
 
-- [Core] Updated `cucumber-core` dependency to [7.0.0-RC1](https://github.com/cucumber/cucumber-jvm/blob/main/CHANGELOG.md)
+- [Core] Updated `cucumber-core` dependency to [7.0.0](https://github.com/cucumber/cucumber-jvm/blob/main/CHANGELOG.md)
 
 ### Deprecated