+++ CHANGELOG.md
@@ -13,6 +13,8 @@ See also the [CHANGELOG](https://github.com/cucumber/cucumber-jvm/blob/master/CH
 
 ### Changed
 
+- [Build] Upgrade Scala to 2.12.14, 2.13.6 and 3.0.2
+
 ### Deprecated
 
 ### Removed