+++ CHANGELOG.md
@@ -11,7 +11,12 @@ See also the [CHANGELOG](https://github.com/cucumber/cucumber-jvm/blob/master/CH
 
 ### Added
 
+- [Scala] Support generic types in `DocStringType`
+
 ### Changed
+
+- [Core] Updated `cucumber-core` dependency to [7.2.1](https://github.com/cucumber/cucumber-jvm/blob/main/CHANGELOG.md)
+- [Scala] Upgrade `jackson-module-scala` to 2.13.1
 
 ### Deprecated